requirements.txt
asgiref==3.7.2
cffi==1.15.1
Django==4.2.3
django-filter==23.2
djangorestframework==3.14.0
Markdown==3.4.3
mysqlclient==2.2.0
Pillow==10.0.0
pycparser==2.21
PyJWT==1.7.1
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
